COLMACICOOP-FINAL/backend/colmacicoop/app/serializers/serviceSerializer.py
from rest_framework import serializers
from app.models import Service

class ServiceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Service
        fields = ['title','content']